Fix/live streaming start from beginning (#12)
* fix: make live streaming playback starts from beginning of buffer

With a HC AAC live streaming asset, the player starts at almost
the tail of the buffer. It seems this behavior comes to enable
seeking. For instance, say we have an asset with 3:00 duration,
the player starts playback from 2:45.

For most of usecases this is bad for UX since:
- playback doesn't start where the user wanted.
- The player often stops playing due to suffring prefetched data.
  (It's seen 15 seconds only)

* fix: revise to seek to the target position on initial play

* fix: ensure the initial seeking happens

* fix: to ensure the initial seek, it needed more heuristic method

Now to achieve it, the result pos must fulfill for 1 second in a row.

* typo

* revert: need to seek when the new url is same with the current